% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hr_ct}
\alias{hr_ct}
\title{CDISC controlled terminology}
\format{
A data frame with columns:
\describe{
  \item{codelist_code}{NCI codelist C-code.}
  \item{codelist_name}{Human-readable codelist name.}
  \item{codelist_submission_value}{Submission value for the codelist.}
  \item{codelist_extensible}{Whether the codelist is extensible.}
  \item{term_code}{NCI term C-code.}
  \item{term_submission_value}{Submission value for the term.}
  \item{term_preferred_name}{NCI preferred term name.}
  \item{term_definition}{Term definition.}
  \item{standard}{SDTM or ADaM.}
}
}
\source{
NCI EVS CDISC Controlled Terminology \url{https://evs.nci.nih.gov/}
}
\usage{
hr_ct
}
\description{
A data frame containing CDISC NCI EVS controlled terminology terms for
SDTM and ADaM. Used by the \pkg{herald} package for terminology validation.
}
\keyword{datasets}
